requirements/base.txt
beautifulsoup4==4.3.2
bpython==0.13
cffi==1.1.1
cryptography==0.9.1
dj-database-url==0.3.0
dj-stripe==0.3.6
Django==1.7
django-allauth==0.20.0
django-allauth-multiple==0.1
django-appconf==1.0.1
django-autofixture==0.10.0
django-avatar==2.0
django-bootstrap-toolkit==2.15.0
django-braces==1.4.0
django-constance==1.0.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-enhanced-cbv==0.1.dev0
django-extensions==1.5.5
django-filter==0.10.0
django-jsonfield==0.9.13
django-model-utils==2.0.3
django-picklefield==0.3.1
django-postgrespool==0.3.0
django-pure-pagination==0.2.1
Django-Select2==4.2.2
django-storages==1.1.8
django-tables2==1.0.4
django-webtest==1.7.8
enum34==1.0.4
gunicorn==19.1.1
idna==2.0
ipaddress==1.0.7
logutils==0.3.3
model-mommy==1.2.4
ndg-httpsclient==0.4.0
oauthlib==0.7.2
Pillow==2.8.2
psycopg2==2.5.4
pyasn1==0.1.7
pycparser==2.13
Pygments==1.6
pyOpenSSL==0.15.1
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
sendgrid==1.4.0
sendgrid-django==1.3.0
six==1.9.0
smtpapi==0.2.0
SQLAlchemy==1.0.5
sqlparse==0.1.15
stripe==1.22.2
waitress==0.8.9
WebOb==1.4.1
WebTest==2.0.18
wheel==0.24.0
whitenoise==1.0.3
